refactor: IssueType issue 패키지로 경로 변경
