+++ src/main/java/dynamicquad/agilehub/issue/controller/response/EpicResponse.java
@@ -1,6 +1,6 @@
 package dynamicquad.agilehub.issue.controller.response;
 
-import dynamicquad.agilehub.issue.controller.request.IssueType;
+import dynamicquad.agilehub.issue.IssueType;
 import dynamicquad.agilehub.issue.domain.epic.Epic;
 import dynamicquad.agilehub.member.dto.AssigneeDto;
 import lombok.Builder;
